refactor: create project structure
