+++ shared/src/commonMain/kotlin/com/akjaw/ai/assistant/shared/App.kt
@@ -5,8 +5,6 @@ package com.akjaw.ai.assistant.shared
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.runtime.*
@@ -16,29 +14,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import io.ktor.client.HttpClient
-import io.ktor.client.plugins.HttpTimeout
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.defaultRequest
-import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logger
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.client.request.headers
-import io.ktor.client.request.post
-import io.ktor.client.request.setBody
-import io.ktor.client.statement.bodyAsText
-import io.ktor.http.ContentType
-import io.ktor.http.HttpStatusCode
-import io.ktor.http.contentType
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
+import com.akjaw.ai.assistant.shared.chat.presentation.ChatScreenStateHolder
+import com.akjaw.ai.assistant.shared.chat.ui.ChatScreen
 import org.jetbrains.compose.resources.ExperimentalResourceApi
 import org.jetbrains.compose.resources.painterResource
-import co.touchlab.kermit.Logger.Companion as KermitLogger
 
 @Composable
 fun App() {
@@ -86,216 +65,5 @@ private fun TypeButton(
     }
 }
 
-val ktorClient = HttpClient {
-    install(Logging) {
-        level = LogLevel.ALL
-        logger = object : Logger {
-            override fun log(message: String) {
-                KermitLogger.i(tag = "Ktor") { message }
-            }
-        }
-    }
-    install(ContentNegotiation) {
-        json()
-    }
-    install(HttpTimeout) {
-        requestTimeoutMillis = 30_000
-        connectTimeoutMillis = 30_000
-        socketTimeoutMillis = 30_000
-    }
-    defaultRequest {
-        url(Endpoints.AddTaskNotion.URL)
-        headers {
-            set("authorization", Endpoints.AddTaskNotion.AUTH)
-        }
-    }
-}
-
-val jsonSerialization = Json {
-    prettyPrint = true
-    isLenient = true
-}
-
-sealed interface ChatMessage {
-
-    val message: String
-
-    data class User(override val message: String) : ChatMessage
-
-    sealed interface Api : ChatMessage {
-
-        data class Success(override val message: String) : Api
-
-        data class Error(override val message: String) : Api
-    }
-}
-
-class AddTask(
-    private val client: HttpClient = ktorClient,
-    private val json: Json = jsonSerialization,
-) {
-
-    @Serializable
-    data class Request(val task: String)
-
-    suspend fun execute(task: String): ChatMessage {
-        val response = client.post {
-            contentType(ContentType.Application.Json)
-            setBody(json.encodeToString(Request(task)))
-        }
-
-        return if (response.status == HttpStatusCode.OK) {
-            ChatMessage.Api.Success(response.bodyAsText())
-        } else {
-            ChatMessage.Api.Error(response.bodyAsText())
-        }
-    }
-}
-
-// TODO add rememberSaveable
-class ChatScreenStateHolder(
-    private val coroutineScope: CoroutineScope,
-    private val addTask: AddTask = AddTask()
-) {
-
-    var userMessage: String by mutableStateOf("")
-        private set
-    val count by derivedStateOf {
-        // TODO count tokens?
-        userMessage.count()
-    }
-    var isLoading: Boolean by mutableStateOf(false)
-        private set
-
-    private val mutableMessages = mutableStateListOf<ChatMessage>()
-    val messages: List<ChatMessage> = mutableMessages
-
-    fun updateUserMessage(message: String) {
-        userMessage = message
-    }
-
-    fun sendMessage() {
-        val message = userMessage
-        mutableMessages.add(ChatMessage.User(message))
-        userMessage = ""
-        isLoading = true
-        coroutineScope.launch {
-            val response = addTask.execute(message)
-            mutableMessages.add(response)
-            isLoading = false
-        }
-    }
-
-    fun retryLastMessage() {
-        val secondToLastMessage =
-            mutableMessages.getOrNull(mutableMessages.lastIndex - 1)
-        if (secondToLastMessage is ChatMessage.User) {
-            mutableMessages.add(secondToLastMessage)
-            isLoading = true
-            coroutineScope.launch {
-                val response = addTask.execute(secondToLastMessage.message)
-                mutableMessages.add(response)
-                isLoading = false
-            }
-        }
-    }
-}
-
-@Composable
-private fun ChatScreen(stateHolder: ChatScreenStateHolder) {
-    // TODO add buttons at the top which changes the "Content"
-    Column(
-        Modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        LazyColumn(
-            modifier = Modifier.weight(1f)
-                .fillMaxWidth()
-                .border(1.dp, Color.Gray, shape = RoundedCornerShape(size = 4.dp))
-                .padding(8.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            item {
-                Spacer(modifier = Modifier.height(8.dp))
-            }
-            itemsIndexed(
-                items = stateHolder.messages,
-                key = { index, _ -> index }
-            ) { _, message ->
-                val background = when (message) {
-                    is ChatMessage.User -> Color.White
-                    is ChatMessage.Api.Error -> Color(0xFFFF7878)
-                    is ChatMessage.Api.Success -> Color(0xFFC9E6C4)
-                }
-                Card(modifier = Modifier.fillMaxWidth(), backgroundColor = background) {
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text(
-                            message.message,
-                            modifier = Modifier.fillMaxWidth().padding(8.dp).weight(1f)
-                        )
-                        if (message is ChatMessage.Api.Error) {
-                            IconButton(
-                                onClick = stateHolder::retryLastMessage,
-                                modifier = Modifier.size(24.dp)
-                            ) {
-                                Image(
-                                    painter = painterResource("refresh.xml"),
-                                    contentDescription = "refresh",
-                                    modifier = Modifier.size(48.dp)
-                                )
-                            }
-                            Spacer(Modifier.width(4.dp))
-                        }
-                    }
-                }
-            }
-        }
-        ChatInput(
-            userMessage = stateHolder.userMessage,
-            setUserMessage = stateHolder::updateUserMessage,
-            count = stateHolder.count,
-            onSend = stateHolder::sendMessage,
-            isEnabled = stateHolder.isLoading.not(),
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-    }
-}
-
-@Composable
-private fun ChatInput(
-    userMessage: String,
-    setUserMessage: (String) -> Unit,
-    count: Int,
-    onSend: () -> Unit,
-    isEnabled: Boolean,
-) {
-    val focusManager = LocalFocusManager.current
-    Row(verticalAlignment = Alignment.CenterVertically) {
-        OutlinedTextField(
-            value = userMessage,
-            onValueChange = setUserMessage,
-            modifier = Modifier.weight(1f),
-            maxLines = 10,
-            enabled = isEnabled
-        )
-        Spacer(modifier = Modifier.width(16.dp))
-        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-            Button(
-                onClick = {
-                    focusManager.clearFocus()
-                    onSend()
-                },
-                enabled = isEnabled
-            ) {
-                Text("Send")
-            }
-            Text(
-                text = count.toString(),
-                style = MaterialTheme.typography.body2.copy(fontSize = 10.sp),
-            )
-        }
-    }
-}
 
 expect fun getPlatformName(): String
